+++ src/events.rs
@@ -93,6 +93,9 @@ pub struct EventSender<E> {
 impl<E> EventSender<E>
 where E: Send + 'static
 {
+    const EVENT_SENDING_ERROR: &'static str =
+        "The associated EventQueue must be alive for sending an event";
+
     fn new(sender: Sender<E>, priority_sender: Sender<E>) -> EventSender<E> {
         EventSender {
             sender,
@@ -105,17 +108,13 @@ where E: Send + 'static
 
     /// Send instantly an event to the event queue.
     pub fn send(&self, event: E) {
-        self.sender
-            .send(event)
-            .expect("The associated EventQueue must be live for sending an event");
+        self.sender.send(event).expect(Self::EVENT_SENDING_ERROR);
     }
 
     /// Send instantly an event that would be process before any other event sent by the send() method.
     /// Successive calls to send_with_priority will maintain the order of arrival.
     pub fn send_with_priority(&self, event: E) {
-        self.priority_sender
-            .send(event)
-            .expect("The associated EventQueue must be live for sending an event");
+        self.priority_sender.send(event).expect(Self::EVENT_SENDING_ERROR);
     }
 
     /// Send a timed event to the [EventQueue].
@@ -137,9 +136,7 @@ where E: Send + 'static
                         return
                     }
                 }
-                sender
-                    .send(event)
-                    .expect("The associated event queue must be live for sending an event");
+                sender.send(event).expect(Self::EVENT_SENDING_ERROR);
             })
             .unwrap();
         self.timer_registry.insert(timer_id, timer_handle);