+++ src/network.rs
@@ -26,20 +26,31 @@ where InMessage: for<'b> Deserialize<'b> + Send + 'static
     Message(Endpoint, InMessage),
 
     /// New endpoint added to a listener.
-    /// It will be sent when a new connection was accepted by the listener.
-    /// This event will be sent only in TCP.
+    /// It is sent when a new connection is accepted by the listener.
+    /// This event will be sent only in connection oriented protocols as TCP.
     AddedEndpoint(Endpoint),
 
     /// A connection lost event.
-    /// This event is only dispatched when a connection is lost. Call to `remove_resource()` will not generate the event.
+    /// This event is only dispatched when a connection is lost.
+    /// Call to `remove_resource()` will not generate the event.
     /// After this event, the resource is considered removed.
     /// A Message event will never be generated after this event.
-    /// This event will be sent only in TCP. Because UDP is not connection oriented, the event can no be detected.
+    /// This event will be sent only in TCP.
+    /// Because UDP is not connection oriented, the event can no be detected.
     RemovedEndpoint(Endpoint),
+
+    /// This event shows that there was a problem during the deserialization of a message.
+    /// The problem is mainly due by a programming issue reading data from an outdated endpoint.
+    /// For example: different protocol version.
+    /// In production it could be that other application is writing in your application port.
+    /// This error means that a message has been lost (the erroneous message),
+    /// but the endpoint remains connected for its usage.
+    DeserializationError(Endpoint),
 }
 
 /// Network allows to manage the network easier.
-/// It is in mainly in charge to transform raw data from the network into message events and vice versa.
+/// It is in mainly in charge to transform raw data from the network into message events
+/// and vice versa.
 pub struct Network {
     network_event_thread: Option<JoinHandle<()>>,
     network_thread_running: Arc<AtomicBool>,
@@ -48,8 +59,11 @@ pub struct Network {
 }
 
 impl<'a> Network {
+    const POISONED_LOCK: &'static str = "This error is shown because other thread has panicked";
+
     /// Creates a new [Network].
-    /// The user must register an event_callback that can be called each time the network generate and [NetEvent]
+    /// The user must register an event_callback that can be called
+    /// each time the network generate and [NetEvent]
     pub fn new<InMessage, C>(event_callback: C) -> Network
     where
         InMessage: for<'b> Deserialize<'b> + Send + 'static,
@@ -109,8 +123,10 @@ impl<'a> Network {
                 log::trace!("Data received from {}, {} bytes", endpoint, data.len());
                 decoding_pool.decode_from(data, endpoint, |decoded_data| {
                     log::trace!("Message received from {}, {} bytes", endpoint, decoded_data.len());
-                    let message: InMessage = bincode::deserialize(decoded_data).unwrap();
-                    event_callback(NetEvent::Message(endpoint, message));
+                    match bincode::deserialize::<InMessage>(decoded_data) {
+                        Ok(message) => event_callback(NetEvent::Message(endpoint, message)),
+                        Err(_) => event_callback(NetEvent::DeserializationError(endpoint)),
+                    }
                 });
             }
             network_adapter::Event::Disconnection => {
@@ -126,8 +142,9 @@ impl<'a> Network {
     /// If the connection can not be performed (e.g. the address is not reached) an error is returned.
     pub fn connect_tcp<A: ToSocketAddrs>(&mut self, addr: A) -> io::Result<Endpoint> {
         let addr = addr.to_socket_addrs().unwrap().next().unwrap();
-        Remote::new_tcp(addr)
-            .map(|remote| self.network_controller.lock().unwrap().add_remote(remote))
+        Remote::new_tcp(addr).map(|remote| {
+            self.network_controller.lock().expect(Self::POISONED_LOCK).add_remote(remote)
+        })
     }
 
     /// Creates a connection to the specific address by UDP.
@@ -135,24 +152,27 @@ impl<'a> Network {
     /// If there is an error during the socket creation, an error will be returned.
     pub fn connect_udp<A: ToSocketAddrs>(&mut self, addr: A) -> io::Result<Endpoint> {
         let addr = addr.to_socket_addrs().unwrap().next().unwrap();
-        Remote::new_udp(addr)
-            .map(|remote| self.network_controller.lock().unwrap().add_remote(remote))
+        Remote::new_udp(addr).map(|remote| {
+            self.network_controller.lock().expect(Self::POISONED_LOCK).add_remote(remote)
+        })
     }
 
     /// Open a port to listen messages from TCP.
     /// If the port can be opened, an resource id identifying the listener is returned along with the local address, or an error if not.
     pub fn listen_tcp<A: ToSocketAddrs>(&mut self, addr: A) -> io::Result<(usize, SocketAddr)> {
         let addr = addr.to_socket_addrs().unwrap().next().unwrap();
-        Listener::new_tcp(addr)
-            .map(|listener| self.network_controller.lock().unwrap().add_listener(listener))
+        Listener::new_tcp(addr).map(|listener| {
+            self.network_controller.lock().expect(Self::POISONED_LOCK).add_listener(listener)
+        })
     }
 
     /// Open a port to listen messages from UDP.
     /// If the port can be opened, an resource id identifying the listener is returned along with the local address, or an error if not.
     pub fn listen_udp<A: ToSocketAddrs>(&mut self, addr: A) -> io::Result<(usize, SocketAddr)> {
         let addr = addr.to_socket_addrs().unwrap().next().unwrap();
-        Listener::new_udp(addr)
-            .map(|listener| self.network_controller.lock().unwrap().add_listener(listener))
+        Listener::new_udp(addr).map(|listener| {
+            self.network_controller.lock().expect(Self::POISONED_LOCK).add_listener(listener)
+        })
     }
 
     /// Open a port to listen messages from UDP in multicast.
@@ -164,9 +184,10 @@ impl<'a> Network {
     ) -> io::Result<(usize, SocketAddr)>
     {
         match addr.to_socket_addrs().unwrap().next().unwrap() {
-            SocketAddr::V4(addr) => Listener::new_udp_multicast(addr)
-                .map(|listener| self.network_controller.lock().unwrap().add_listener(listener)),
-            _ => panic!("listening for udp multicast is only supported for ipv4 addresses"),
+            SocketAddr::V4(addr) => Listener::new_udp_multicast(addr).map(|listener| {
+                self.network_controller.lock().expect(Self::POISONED_LOCK).add_listener(listener)
+            }),
+            _ => panic!("Listening for udp multicast is only supported for ipv4 addresses"),
         }
     }
 
@@ -177,59 +198,54 @@ impl<'a> Network {
     /// Note: Udp endpoints generated by a UdpListener shared the resource, the own UdpListener.
     /// This means that there is no resource to remove as the TCP case.
     pub fn remove_resource(&mut self, resource_id: usize) -> Option<()> {
-        self.network_controller.lock().unwrap().remove_resource(resource_id)
+        self.network_controller.lock().expect(Self::POISONED_LOCK).remove_resource(resource_id)
     }
 
     /// Request a local address of a resource.
     /// Returns `None` if the endpoint id not exists.
     /// Note: Udp endpoints generated by a UdpListener shared the resource.
     pub fn local_address(&self, resource_id: usize) -> Option<SocketAddr> {
-        self.network_controller.lock().unwrap().local_address(resource_id)
+        self.network_controller.lock().expect(Self::POISONED_LOCK).local_address(resource_id)
     }
 
     /// Serialize and send the message thought the connection represented by the given endpoint.
     /// If the same message should be sent to different endpoints, use `send_all()` to better performance.
-    /// Returns an error if there is an error while sending the message, the endpoint does not exists, or if it is not valid.
-    pub fn send<OutMessage>(&mut self, endpoint: Endpoint, message: OutMessage) -> io::Result<()>
+    /// The funcion panics if some of endpoints do not exists.
+    /// If the endpoint disconnects during the sending, a RemoveEndpoint is generated.
+    pub fn send<OutMessage>(&mut self, endpoint: Endpoint, message: OutMessage)
     where OutMessage: Serialize {
         self.prepare_output_message(message);
-        let result = self.network_controller.lock().unwrap().send(endpoint, &self.output_buffer);
+        self.network_controller
+            .lock()
+            .expect(Self::POISONED_LOCK)
+            .send(endpoint, &self.output_buffer);
+
         self.output_buffer.clear();
-        if result.is_ok() {
-            log::trace!("Message sent to {}", endpoint);
-        }
-        result
+        log::trace!("Message sent to {}", endpoint);
     }
 
     /// Serialize and send the message thought the connections represented by the given endpoints.
     /// When there are severals endpoints to send the data, this function is faster than consecutive calls to `send()`
     /// since the serialization is only performed one time for all endpoints.
-    /// A list of erroneous endpoints along with their errors is returned if there was a problem with some messages sent.
-    /// An error sending a message by an endpoint do not avoid sending the message to the rest of endpoints.
+    /// The funcion panics if some of endpoints do not exists.
+    /// If the protocol is UDP, the function panics if the message size is higher than MTU.
+    /// If the endpoint disconnects during the sending, a RemoveEndpoint is generated.
     pub fn send_all<'b, OutMessage>(
         &mut self,
         endpoints: impl IntoIterator<Item = &'b Endpoint>,
         message: OutMessage,
-    ) -> Result<(), Vec<(Endpoint, io::Error)>>
-    where
+    ) where
         OutMessage: Serialize,
     {
         self.prepare_output_message(message);
-        let mut errors = Vec::new();
-        let mut controller = self.network_controller.lock().unwrap();
+        let mut controller = self.network_controller.lock().expect(Self::POISONED_LOCK);
+
         for endpoint in endpoints {
-            match controller.send(*endpoint, &self.output_buffer) {
-                Ok(_) => log::trace!("Message sent to {}", endpoint),
-                Err(err) => errors.push((*endpoint, err)),
-            }
+            controller.send(*endpoint, &self.output_buffer);
+            log::trace!("Message sent to {}", endpoint);
         }
+
         self.output_buffer.clear();
-        if errors.is_empty() {
-            Ok(())
-        }
-        else {
-            Err(errors)
-        }
     }
 
     fn prepare_output_message<OutMessage>(&mut self, message: OutMessage)
@@ -243,6 +259,10 @@ impl<'a> Network {
 impl Drop for Network {
     fn drop(&mut self) {
         self.network_thread_running.store(false, Ordering::Relaxed);
-        self.network_event_thread.take().unwrap().join().unwrap();
+        self.network_event_thread
+            .take()
+            .expect("Always exists")
+            .join()
+            .expect("This error is shown because other thread has panicked");
     }
 }