+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Release 0.5.0
 - Renamed: `receive_event_timeout` to `receive_timeout`.
 - Renaded: `NetworkManager` to `Network`.
+- Fixed Reset-by-peer errors.
+  Now it generates a RemoveEndpoint event.
+- Added DeserializationError Event that is generated when there is a problem at deserializing.
+- Removed Result from `send()` and `send_all()` functions.
+  Now if some error happens during sending, a RemovedEndpoint will be generated (only in Tcp).
 
 ## Release 0.4.6
 - Fixed lost decoding memory at disconnection.